// frontend/src/app/faculty/+state/enums/faculty-status.enum.ts
export enum FacultyStatusEnum {
  pending = 'pending',
  loading = 'loading',
  loadSuccess = 'loadSuccess',
  loadError = 'loadError',
  loadDetailsSuccess = 'loadDetailsSuccess',
  loadDetailsFailure = 'loadDetailsFailure',
  createSuccess = 'createSuccess',
  createFailure = 'createFailure',
  updateSuccess = 'updateSuccess',
  updateFailure = 'updateFailure',
  deleteSuccess = 'deleteSuccess',
  deleteFailure = 'deleteFailure',
  approveSuccess = 'approveSuccess',
  approveFailure = 'approveFailure',
}